add a button to delete the certificate then modify the nit attribute in the person model

+++ sistema_laboratorio/app/cliente/models.py
@@ -6,7 +6,7 @@ class Cliente(models.Model):
 	Nombre=models.CharField(max_length=150)
 	Direccion =models.CharField(max_length=100)
 	Telefono = models.PositiveIntegerField(blank=True, null=True, help_text="Opcional")
-	Nit = models.PositiveIntegerField(blank=True, null=True,help_text="Opcional")
+	Nit = models.BigIntegerField(blank=True, null=True,help_text="Opcional")
 	Particular = models.BooleanField()
 	estado=models.BooleanField(default=True)
 	fecha_registro = models.DateTimeField(auto_now_add=True)
